requirements.txt
openai>=1.0
anthropic>=0.30
google-genai>=1.0
httpx
pyyaml
python-dotenv
pandas
tenacity
